+++ src/Gate.IO.Api/Earn/Enums/GateEarnStakingAssetStatus.cs
@@ -1,0 +1,17 @@
+namespace Gate.IO.Api.Earn;
+
+/// <summary>
+/// Staking asset project status
+/// </summary>
+public enum GateEarnStakingAssetStatus : byte
+{
+    /// <summary>
+    /// Currency project is listed
+    /// </summary>
+    Listed = 1,
+
+    /// <summary>
+    /// Currency project is delisted
+    /// </summary>
+    Delisted = 2,
+}