+++ src/Gate.IO.Api/Earn/Enums/GateEarnStakingAwardStatus.cs
@@ -1,0 +1,12 @@
+namespace Gate.IO.Api.Earn;
+
+/// <summary>
+/// Staking award status
+/// </summary>
+public enum GateEarnStakingAwardStatus : byte
+{
+    /// <summary>
+    /// Award distributed successfully
+    /// </summary>
+    Success = 4,
+}